// constants/enums.ts
export enum Errors {
    ERROR_VALIDATION = 'ERROR VALIDATION',
    ERROR_NO_DATA = 'ERROR NO DATA',
    INVALID_ID_FORMAT = 'INVALID ID FORMAT',
    ERROR_UNKNOWN = 'ERROR UNKNOWN',
    INTERNAL_SERVER_ERROR = 'INTERNAL SERVER ERROR',
    USER_NOT_FOUND = 'USER NOT FOUND',
    ERROR_REPEATED = 'ERROR REPEATED'
}

export enum SucessResponse{
    SUCCESS = 'SUCCESS'
}